Centralize publisher API creation
